remove allowlist, only use denylist, enable newly passing tests
